Implement GetBudgets and GetRunway
It turns out these are the questions I care most about. I deleted a lot
of planning tests, because I don't actually care about fine-grained
details of priority allocation (and I think the test coverage around
saving-schedules is sufficient for that anyway).

What matters is I need to know whether I'm over-or-under-budget for a
particular target on any given week (so I know how much I can spend at
the grocery store, or whether I have budget for a special event), and
what my time horizon for running out of savings is so I can track if
it's creeping closer.