Update GetBalances to accept date argument
Generally all operations should accept an argument specifying what day
to consider "today". GetBalances was somewhat dangerous in that it
always returned the balances over the entire event stream, but it was
used in composite operations with operations that respected a given
date. In practice this could have resulted in incorrect or meaningless
results if dates other than the end of the event stream were ever
provided.

In particular, I expect to need to take advantage of the date parameter
to calculate runway trends over time.